+++ examples/01_plotting/plot_haxby_masks.py
@@ -22,7 +22,7 @@ from nilearn import image
 func_filename = haxby_dataset.func[0]
 mean_img = image.mean_img(func_filename)
 
-z_slice = -24
+z_slice = -14
 from nilearn.image.resampling import coord_transform
 affine = mean_img.affine
 _, _, k_slice = coord_transform(0, 0, z_slice,